Add better profiling functionality in CudaRTCFunction
CUPTI is used to query gpu performance counters/events and compute
metrics such as occupancy. Not all events can be measured in one run, as
such for a particular metric multiple kernel launches might be required.

This commit only includes a set of hardcoded metrics (later on these
should be selectable).